fixed GPIOWrite function and minor graphic bugs

+++ IN3_FW/motherBoard/IN3_FW/include/main.h
@@ -341,7 +341,7 @@ void drawBack();
 void drawActuatorsSeparators();
 void drawStop();
 void drawHelpMessage(byte UI_language);
-void printLoadingTemperatureBar(int UI_desiredControlTemperature);
+void printLoadingTemperatureBar(double UI_desiredControlTemperature);
 void printLoadingHumidityBar(int UI_desiredControlHumidity);
 void blinkGoBackMessage();
 bool ongoingAlarms();
